-Внедрение statRepo через DependencyInjection -Переработка кода StatValue (теперь класс называется StatInfo) и, как следствие, StatsCalc и наследников.

+++ src/main/java/com/stsetsevich/smartpoker/engine/AddOrUpdatePlayer.java
@@ -44,24 +44,26 @@ public class AddOrUpdatePlayer {
 
     }
     public void updatePlayerIfNeed(Player player) {
-        java.util.Date lastUpdate = player.getDateUpdate();
-        java.util.Date today = new Date(System.currentTimeMillis());
-        double totalHands = player.getTotalHands();
+        if (!player.getNickname().equals("Empty Seat")) {
+            java.util.Date lastUpdate = player.getDateUpdate();
+            java.util.Date today = new Date(System.currentTimeMillis());
+            double totalHands = player.getTotalHands();
 
-        long updateOld = -100;
-        try {
-            updateOld = (today.getTime() - lastUpdate.getTime()) / (24 * 60 * 60 * 1000);
-        } catch (Exception exception) {
+            long updateOld = -100;
+            try {
+                updateOld = (today.getTime() - lastUpdate.getTime()) / (24 * 60 * 60 * 1000);
+            } catch (Exception exception) {
 
-        }
+            }
 
-        if (updateOld < -99 ||
-                (updateOld >= 1 && totalHands <= 5000) ||
-                (updateOld >= 5 && totalHands <= 15000) ||
-                (updateOld >= 14 && totalHands <= 50000) ||
-                (updateOld >= 30)) {
-            playerRepo.delete(player);
-            tryAddNewPlayer(player.getNickname());
+            if (updateOld < -99 ||
+                    (updateOld >= 1 && totalHands <= 5000) ||
+                    (updateOld >= 5 && totalHands <= 15000) ||
+                    (updateOld >= 14 && totalHands <= 50000) ||
+                    (updateOld >= 30)) {
+                playerRepo.delete(player);
+                tryAddNewPlayer(player.getNickname());
+            }
         }
     }
 }
